Move motor module under hoverboard module.
