+++ cross/hoverkite-firmware/src/hoverboard/motor.rs
